Create understanding_of_EmonLib.cpp
wanting to understand how the calcVI and calcIRMS work. Redoing to eliminate arduino with ADS1115. Just need to know the math to work on the signals. They are doing something in hear as a low pass filter and a high pass filter via software. Just trying to figure out how to recreate that in python possibly.