Create rewrite_of_Arduino_DRO_V5_10
Has a lot of functions we can eliminate since we aren't using the W axis. 
Mainly, I want to understand how they are capturing the tachometer data, processing the data, and producing it. If we can figure out how to do this with an raspberry pi script it would be great!
Also, if you see a method of reading the position that we can convert with their counting technique we can capture that as well.